# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: optparts.kz
  version: 1.0.0
  description: |
    JWT-based authentication API with user management, product catalog, shopping cart, orders, and delivery addresses.

    ## Authentication
    Most endpoints require a Bearer JWT token. Obtain one via `/auth/login` or `/auth/register`, then pass it as:
    ```
    Authorization: Bearer <token>
    ```

    ## Roles
    | Role | Value |
    |------|-------|
    | User | `user` |
    | Manager | `manager` |
    | Admin | `admin` |
servers:
  - url: /api
    description: Local server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: c1d2e3f4-a5b6-7890-cdef-123456789012
        name:
          type: string
          example: Electronics
        description:
          type: string
          nullable: true
          example: Electronic devices and accessories
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ─── Brands ──────────────────────────────────────────────────────────────
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: b1c2d3e4-f5a6-7890-bcde-123456789012
        name:
          type: string
          example: SoundMax
        description:
          type: string
          nullable: true
          example: Premium audio equipment
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ─── Products ────────────────────────────────────────────────────────────
...
    Address:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Home
        street:
          type: string
          example: 123 Abay Ave
        city:
          type: string
          example: Almaty
        postcode:
          type: string
          example: "050000"
        main:
          type: boolean
          example: true
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ─── Orders ──────────────────────────────────────────────────────────────
...
    ValidationError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: array
          items:
            type: string
          example: ["email must be an email", "password must be at least 8 characters"]
        error:
          type: string
          example: Bad Request
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Access denied. Please log in.
        error:
          type: string
          example: Unauthorized
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 403
        message:
          type: string
          example: Insufficient permissions.
        error:
          type: string
          example: Forbidden
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Resource not found.
        error:
          type: string
          example: Not Found
security:
  - bearerAuth: []
paths:
...
  /categories:
    get:
      tags: [Categories]
      summary: List all categories
      security: []
      responses:
        '200':
          description: Array of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      tags: [Categories]
      summary: Create category (admin / manager)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: Electronics
                description:
                  type: string
                  example: Electronic devices and accessories
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
  /categories/{id}:
    get:
      tags: [Categories]
      summary: Get category by ID
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
    put:
      tags: [Categories]
      summary: Update category (admin / manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
    delete:
      tags: [Categories]
      summary: Delete category (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

  # ═══════════════════════════════════════════════════════════════════════════
  # BRANDS
  # ═══════════════════════════════════════════════════════════════════════════
  /brands:
    get:
      tags: [Brands]
      summary: List all brands
      security: []
      responses:
        '200':
          description: Array of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brand'
    post:
      tags: [Brands]
      summary: Create brand (admin / manager)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: SoundMax
                description:
                  type: string
                  example: Premium audio equipment
      responses:
        '201':
          description: Brand created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
  /brands/{id}:
    get:
      tags: [Brands]
      summary: Get brand by ID
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Brand
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
    put:
      tags: [Brands]
      summary: Update brand (admin / manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Brand updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
    delete:
      tags: [Brands]
      summary: Delete brand (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

  # ═══════════════════════════════════════════════════════════════════════════
  # PRODUCTS
  # ═══════════════════════════════════════════════════════════════════════════
...
  /addresses:
    get:
      tags: [Addresses]
      summary: List all addresses for current user (main first)
      responses:
        '200':
          description: Array of addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
    post:
      tags: [Addresses]
      summary: Create address — setting main:true auto-unsets previous main
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, street, city, postcode]
              properties:
                title:
                  type: string
                  example: Home
                street:
                  type: string
                  example: 123 Abay Ave
                city:
                  type: string
                  example: Almaty
                postcode:
                  type: string
                  example: "050000"
                main:
                  type: boolean
                  example: false
      responses:
        '201':
          description: Address created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
  /addresses/{id}:
    get:
      tags: [Addresses]
      summary: Get address by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Address belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
    put:
      tags: [Addresses]
      summary: Update address fields
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                street:
                  type: string
                city:
                  type: string
                postcode:
                  type: string
                main:
                  type: boolean
      responses:
        '200':
          description: Address updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
    delete:
      tags: [Addresses]
      summary: Delete address
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /addresses/{id}/main:
    patch:
      tags: [Addresses]
      summary: Set address as main — auto-unsets previous main
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Address is now main
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

  # ═══════════════════════════════════════════════════════════════════════════
  # ORDERS
  # ═══════════════════════════════════════════════════════════════════════════
...